Scale pseudocounts for `dms2_prefs --method ratio`
@djpl had noticed that you could get strange results
when inferring preferences using `--method ratio` when
the depth was very unequal across samples and the counts
were low.

It turns out that this is because the pseudocount was
not scaled proportionally to depth, which led to systematically
different frequency estimates for samples of different depth.

This is now fixed by scaling the pseudocount by depth.